Convert back to double for bounds

+++ src/bound.h
@@ -2,7 +2,7 @@
 #define BOUND_H
 
 double hermite_constant(const int dim);
-long double volL(const Vector2D *Bs, const int dim);
-long double lambda_1(const Vector2D *Bs, const int dim);
+double volL(const Vector2D *Bs, const int dim);
+double lambda_1(const Vector2D *Bs, const int dim);
 
 #endif // BOUND_H
