+++ src/schorr_euchner.c
@@ -31,12 +31,12 @@ void init_var(int dim, GS_Info *gs_info, double **p, int **v, double **c, int **
     (*v)[0] = 1;
 }
 
-double schorr_euchner(const int dim, GS_Info *gs_info, long double R) {
+double schorr_euchner(const int dim, GS_Info *gs_info, double R) {
     double *p, *c;
     int *v, *w;
     init_var(dim, gs_info, &p, &v, &c, &w);
 
-    long double R_2 = R * R;
+    double R_2 = R * R;
     int k = 0;
     int last_non_zero = 0;
